fix(range-info-bar): range level value z-index
